[CODING] fix <Tab> unusable error [fix]
I used to set <C-i> as the keybinding to manually trigger completion,
however, since <C-i> is sent out as ^I in terminal and it is the code
for <Tab>, we do not have <Tab> anymore.

The problem here is that since we did not set keybindings for <Tab>
but <C-i>, we cannot find this issue via check keybinding.